% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series_system.R
\name{md_exp_series_system_bernoulli_cand_C1_C2_C3}
\alias{md_exp_series_system_bernoulli_cand_C1_C2_C3}
\title{Generate exponential series data with right-censoring and masked
component failure satisfying conditions C1, C2, and C3.}
\usage{
md_exp_series_system_bernoulli_cand_C1_C2_C3(n, rates, p = 0, tau = Inf)
}
\arguments{
\item{n}{number of observations}

\item{rates}{vector of component failure rates}

\item{p}{numeric, \code{p[i]} is the probability that each of the
non-failed components in the i-th series system
appears in the candidate set. Defaults to rep(0,n).
If p\link{i} = 0, then only the failed component is in the
candidate set (no masking). If p\link{i} = 1, then all
all components are in the candidate set, in which case
the candidate set conveys no information about the
(latent) components (the lifetime of the system still
conveys information about the components, but if that's
all the sample has, then the MLE is not
identifiable).}

\item{tau}{numeric, right-censoring time, defaults to rep(Inf,n).
If tau\link{i} = Inf, then the i-th system is not
right-censored.}
}
\description{
Generate exponential series data with right-censoring and masked
component failure satisfying conditions C1, C2, and C3.
}
