% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_exp_series_system.R
\name{md_score_exp_series_C1_C2_C3}
\alias{md_score_exp_series_C1_C2_C3}
\title{Generates a score function for an exponential series system (or
competing risks) with respect to parameter \code{theta} for masked component
failure (or masked competing risks) with candidate sets (risks) that
approximately satisfy conditions C1, C2, and C3.}
\usage{
md_score_exp_series_C1_C2_C3(md, options = list(), ...)
}
\arguments{
\item{md}{right-censored failure-time data with masked competing risks}

\item{options}{list of options
\itemize{
\item \code{set.var} prefix of Boolean matrix encoding of candidate sets, defaults
to \code{x}, e.g., \verb{x1,...,xm}.
\item \code{sys.var} system lifetime (optionally right-censored) column name,
defaults to \code{t}
\item \code{delta.var} right-censoring indicator column name. If \code{NULL}, then
no right-censoring is assumed. If a system lifetime is
right-censored (\emph{not} observed), the right-censoring
indicator is \code{TRUE}, otherwise it is \code{FALSE}.
}}

\item{...}{pass additional arguments to \code{options}}
}
\value{
score function of type \code{R^m -> R}
}
\description{
Generates a score function for an exponential series system (or
competing risks) with respect to parameter \code{theta} for masked component
failure (or masked competing risks) with candidate sets (risks) that
approximately satisfy conditions C1, C2, and C3.
}
